Django Task 4 & Task 5 completed

+++ django/students/views.py
@@ -1,5 +1,7 @@
 from django.shortcuts import render, get_object_or_404
 from django.shortcuts import redirect, HttpResponseRedirect
+from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
+from django.contrib import messages
 from courses.models import Course
 from .models import Student
 from .forms import StudentForm
@@ -16,7 +18,15 @@ def students_list(request):
     else:
         course = get_object_or_404(Course, pk=pk)
         students = Student.objects.filter(courses=course).order_by('name')
-    return render(request, 'students/students_list.html', {'students': students})
+    paginator = Paginator(students, 2)
+    page = request.GET.get('page')
+    try:
+        students_page = paginator.page(page)
+    except PageNotAnInteger:
+        students_page = paginator.page(1)
+    except EmptyPage:
+        students_page = paginator.page(paginator.num_pages)
+    return render(request, 'students/students_list.html', {'students': students_page})
 
 
 def student_details(request, pk):
@@ -30,11 +40,14 @@ def student_new(request):
         form = StudentForm(request.POST)
         if (form.is_valid()):
             student = form.save()
+            name = student.name
+            surname = student.surname
             next = request.POST.get('next', '/')
+            messages.success(request, "Студент " + name + " " + surname + " был успешно добавлен")
             return HttpResponseRedirect(next)
     else:
         course_id = request.GET.get('course_id', '-1')
-        if '-' in course_id or course_id=='':
+        if '-' in course_id or course_id == '':
             form = StudentForm()
         else:
             course = get_object_or_404(Course, pk=course_id)
@@ -49,15 +62,27 @@ def student_edit(request, pk):
         form = StudentForm(request.POST, instance=student)
         if (form.is_valid()):
             student = form.save()
+            name = student.name
+            surname = student.surname
             next = request.POST.get('next', '/')
+            messages.success(request, "Студент " + name + " " + surname + " был успешно удален")
             return HttpResponseRedirect(next)
     else:
         form = StudentForm(instance=student)
     return render(request, 'students/student_edit.html', {'form': form})
 
+
 @login_required
-def student_remove(request,pk):
-    student=get_object_or_404(Student,pk=pk)
-    student.delete()
-    next = request.GET.get('next', '/')
-    return HttpResponseRedirect(next)
+def student_remove(request, pk):
+    student = get_object_or_404(Student, pk=pk)
+    if (request.method == "POST"):
+        name = student.name
+        surname = student.surname
+        student.delete()
+        next = request.GET.get('next', '/')
+        messages.success(request, "Студент " + name + " " + surname + " был успешно удален")
+        return HttpResponseRedirect(next)
+    else:
+        name = student.name
+        surname = student.surname
+    return render(request, 'students/remove.html', {'name': name, 'surname': surname, 'pk': pk})
